Update the scope of flow control.
